refactor(eager): split _PropTorchEager into core + 2 mixins
_PropTorchEager had grown to ~914 lines / 29 methods mixing three
concerns.  Extract the two self-contained features into mixins (same
self.* methods, resolved via MRO — behaviour-preserving):

- _CustomGradientMixin (_torch_eager_custom_grad.py): register_gradient /
  imaging / the hook-based adjoint-state forward + _GradBridgeFn.
- _EagerBoundarySavingMixin (_torch_eager_boundary.py): reverse-driver
  dispatch, per-rollout init, the boundary-saving time loop.

The core _PropTorchEager keeps construction, stepping/compile, snapshots,
ckpt/full rollout and forward (~360 lines).  Class bases append the
mixins after the concrete bases so __init__/super() are unchanged.

Verified: import resolves all methods via MRO, eager BS pytest 27/27,
4-way eB/eF cosine 1.0 across all equations.
